% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaldist.R
\name{pdfnormaldist}
\alias{pdfnormaldist}
\title{normal_distribution}
\usage{
pdfnormaldist(x, a, b)
}
\description{
This function calculates the value of pdf of normal distribution.
}
\examples{

pdfnormaldist(2,30,9)
}
